refactor(api): move findPendingInvitations controller to src/team
